+++ src/service/translation/__init__.py
@@ -1,0 +1,20 @@
+from translator import Translator
+from deepl_translator import DeepLTranslator
+from google_translator import GoogleTranslator
+from gpt_translator import GPTTranslator
+from .srt import srt_sentense_merge, srt_to_text
+
+
+def get_translator(translate_vendor, api_key=None, proxies=None):
+    if translate_vendor == "google":
+        return GoogleTranslator(proxy=proxies)
+    elif translate_vendor == "deepl":
+        if not api_key:
+            raise ValueError("Missing translate key for DeepL.")
+        return DeepLTranslator(key=api_key)
+    elif "gpt" in translate_vendor:
+        if not api_key:
+            raise ValueError("Missing translate key for GPT.")
+        return GPTTranslator(api_key=api_key, model_name=translate_vendor, proxies=proxies)
+    else:
+        raise ValueError("Unknown translation vendor.")
