+++ src/service/translation/srt.py
@@ -1,6 +1,5 @@
 import copy
 import re
-
 import srt
